Make delegate and dataSource injectable
