+++ TableViewKit/TableViewManager.swift
@@ -11,43 +11,39 @@ open class TableViewManager {
     open let tableView: UITableView
 
     /// An array of sections
-    open var sections: ObservableArray<Section>
+	open var sections: ObservableArray<Section>
 
-	open var dataSource: TableViewKitDataSource?
-
-	open var delegate: TableViewKitDelegate?
+	open var dataSourceClass: TableViewKitDataSourceType.Type { return TableViewKitDataSource.self }
+	open var delegateClass: TableViewKitDelegateType.Type { return TableViewKitDelegate.self }
 
     open var animation: UITableViewRowAnimation = .automatic
 
+	public var dataSource: TableViewKitDataSourceType?
+	public var delegate: TableViewKitDelegateType?
+
     var reusableIdentifiers: Set<String> = []
-
-    /// Initialize a `TableViewManager` with a `tableView`.
-    ///
-    /// - parameter tableView: A `tableView` that will be controlled by the `TableViewManager`
-    public init(tableView: UITableView) {
-        self.tableView = tableView
-        self.sections = []
-		self.dataSource = TableViewKitDataSource(manager: self)
-		self.delegate = TableViewKitDelegate(manager: self)
-        self.tableView.dataSource = self.dataSource
-        self.tableView.delegate = self.delegate
-        self.setupSections()
-
-    }
 
     /// Initialize a `TableViewManager` with a `tableView` and an initial array of sections
     ///
     /// - parameter tableView: A `tableView` that will be controlled by the `TableViewManager`
     /// - parameter sections: An array of sections
-    public init(tableView: UITableView, sections: [Section]) {
+    public init(tableView: UITableView, sections: [Section] = []) {
         self.tableView = tableView
         self.sections = ObservableArray(array: sections)
-		self.dataSource = TableViewKitDataSource(manager: self)
-		self.delegate = TableViewKitDelegate(manager: self)
-		self.tableView.dataSource = self.dataSource
-		self.tableView.delegate = self.delegate
+		self.setupDelegate()
+		self.setupDataSource()
         self.setupSections()
     }
+
+	private func setupDelegate() {
+		self.delegate = delegateClass.init(manager: self)
+		self.tableView.delegate = self.delegate
+	}
+
+	private func setupDataSource() {
+		self.dataSource = dataSourceClass.init(manager: self)
+		self.tableView.dataSource = self.dataSource
+	}
 
     private func setupSections() {
         sections.forEach { section in
@@ -102,175 +98,8 @@ extension TableViewManager {
             reusableIdentifiers.insert(type.reusableIdentifier)
         }
     }
-}
 
-open class TableViewKitDataSource: NSObject, UITableViewDataSource {
-
-	open unowned var manager: TableViewManager
-	open var sections: ObservableArray<Section> { return manager.sections }
-
-	public init(manager: TableViewManager) {
-		self.manager = manager
-	}
-
-
-    /// Implementation of UITableViewDataSource
-    open func numberOfSections(in tableView: UITableView) -> Int {
-        return sections.count
-    }
-
-    /// Implementation of UITableViewDataSource
-    open func tableView(_ tableView: UITableView, numberOfRowsInSection sectionIndex: Int) -> Int {
-        let section = sections[sectionIndex]
-        return section.items.count
-    }
-
-    /// Implementation of UITableViewDataSource
-    open func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let currentItem = item(at: indexPath)
-        let drawer = type(of: currentItem).drawer
-
-        let cell = drawer.cell(in: manager, with: currentItem, for: indexPath)
-        drawer.draw(cell, with: currentItem)
-
-        return cell
-    }
-
-    /// Implementation of UITableViewDataSource
-    open func tableView(_ tableView: UITableView, titleForHeaderInSection section: Int) -> String? {
-        return title(for: {$0.header}, inSection: section)
-    }
-
-    /// Implementation of UITableViewDataSource
-    open func tableView(_ tableView: UITableView, titleForFooterInSection section: Int) -> String? {
-        return title(for: {$0.footer}, inSection: section)
-    }
-
-    /// Implementation of UITableViewDataSource
-    open func tableView(_ tableView: UITableView, commit editingStyle: UITableViewCellEditingStyle, forRowAt indexPath: IndexPath) {
-        // Intentionally blank. Required to use UITableViewRowActions
-    }
-
-	/// Implementation of UITableViewDataSource
-	open func tableView(_ tableView: UITableView, canEditRowAt indexPath: IndexPath) -> Bool {
-		return item(at: indexPath) is Editable
-	}
-
-	fileprivate func title(for key: (Section) -> HeaderFooterView, inSection section: Int) -> String? {
-		if case .title(let value) = key(sections[section]) {
-			return value
-		}
-		return nil
-	}
-
-	fileprivate func item(at indexPath: IndexPath) -> Item {
+	func item(at indexPath: IndexPath) -> Item {
 		return sections[indexPath.section].items[indexPath.row]
 	}
 }
-
-open class TableViewKitDelegate: NSObject, UITableViewDelegate {
-
-	open unowned var manager: TableViewManager
-	open var sections: ObservableArray<Section> { return manager.sections }
-
-	public init(manager: TableViewManager) {
-		self.manager = manager
-	}
-
-    /// Implementation of UITableViewDelegate
-    open func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        guard let currentItem = item(at: indexPath) as? Selectable else { return }
-        currentItem.didSelect()
-    }
-
-    /// Implementation of UITableViewDelegate
-    open func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
-        return height(at: indexPath) ?? tableView.rowHeight
-    }
-
-    /// Implementation of UITableViewDelegate
-    open func tableView(_ tableView: UITableView, heightForHeaderInSection section: Int) -> CGFloat {
-        return height(for: {$0.header}, inSection: section) ?? tableView.sectionHeaderHeight
-    }
-
-    /// Implementation of UITableViewDelegate
-    open func tableView(_ tableView: UITableView, heightForFooterInSection section: Int) -> CGFloat {
-        return height(for: {$0.footer}, inSection: section) ?? tableView.sectionFooterHeight
-    }
-
-    /// Implementation of UITableViewDelegate
-    open func tableView(_ tableView: UITableView, estimatedHeightForRowAt indexPath: IndexPath) -> CGFloat {
-        return estimatedHeight(at: indexPath) ?? tableView.estimatedRowHeight
-    }
-
-    /// Implementation of UITableViewDelegate
-    open func tableView(_ tableView: UITableView, estimatedHeightForHeaderInSection section: Int) -> CGFloat {
-        return estimatedHeight(for: {$0.header}, inSection: section) ?? tableView.estimatedSectionHeaderHeight
-    }
-
-    /// Implementation of UITableViewDelegate
-    open func tableView(_ tableView: UITableView, estimatedHeightForFooterInSection section: Int) -> CGFloat {
-        return estimatedHeight(for: {$0.footer}, inSection: section) ?? tableView.estimatedSectionHeaderHeight
-    }
-
-    /// Implementation of UITableViewDelegate
-    open func tableView(_ tableView: UITableView, viewForHeaderInSection section: Int) -> UIView? {
-        return view(for: {$0.header}, inSection: section)
-    }
-
-    /// Implementation of UITableViewDelegate
-    open func tableView(_ tableView: UITableView, viewForFooterInSection section: Int) -> UIView? {
-        return view(for: {$0.footer}, inSection: section)
-    }
-
-    /// Implementation of UITableViewDelegate
-    open func tableView(_ tableView: UITableView, editActionsForRowAt indexPath: IndexPath) -> [UITableViewRowAction]? {
-        guard let item = item(at: indexPath) as? Editable else { return nil }
-        return item.actions
-    }
-
-	fileprivate func item(at indexPath: IndexPath) -> Item {
-		return sections[indexPath.section].items[indexPath.row]
-	}
-
-	fileprivate func view(for key: (Section) -> HeaderFooterView, inSection section: Int) -> UIView? {
-		guard case .view(let item) = key(sections[section]) else { return nil }
-
-		let drawer = type(of: item).drawer
-		let view = drawer.view(in: manager, with: item)
-		drawer.draw(view, with: item)
-
-		return view
-	}
-
-	fileprivate func estimatedHeight(for key: (Section) -> HeaderFooterView, inSection section: Int) -> CGFloat? {
-		let item = key(sections[section])
-		switch item {
-		case .view(let view):
-			guard let height = view.height else { return nil }
-			return height.estimated
-		case .title(_):
-			return 1.0
-		default:
-			return nil
-		}
-	}
-
-	fileprivate func estimatedHeight(at indexPath: IndexPath) -> CGFloat? {
-		guard let height = item(at: indexPath).height else { return nil }
-		return height.estimated
-	}
-
-	fileprivate func height(for key: (Section) -> HeaderFooterView, inSection section: Int) -> CGFloat? {
-		guard case .view(let view) = key(sections[section]), let value = view.height
-			else { return nil }
-		return value.height
-	}
-
-	fileprivate func height(at indexPath: IndexPath) -> CGFloat? {
-		guard let value = item(at: indexPath).height else { return nil }
-		return value.height
-	}
-
-}
-
